fixed time and form errors on projects page and addProject backend filr

+++ freelancelotDB.sql
@@ -44,7 +44,7 @@ ptitle varchar(255) not null,
 pdescription text(65535) not null,
 createTime datetime not null,
 payStatus varchar(255) not null,
-amount float(5,2) not null,
+amount float(9,2) not null,
 pdifficulty varchar(255) not null,
 foreign key (organID) references organizations(organID)
 );
